Move pojo class to inlong-common.
