+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/repository/DataProxyConfigRepository.java
@@ -33,11 +33,13 @@ import org.apache.inlong.commons.pojo.dataproxy.CacheClusterSetObject;
 import org.apache.inlong.commons.pojo.dataproxy.CacheTopicObject;
 import org.apache.inlong.commons.pojo.dataproxy.DataProxyCluster;
 import org.apache.inlong.commons.pojo.dataproxy.DataProxyConfigResponse;
+import org.apache.inlong.commons.pojo.dataproxy.IRepository;
 import org.apache.inlong.commons.pojo.dataproxy.InLongIdObject;
 import org.apache.inlong.commons.pojo.dataproxy.ProxyChannel;
 import org.apache.inlong.commons.pojo.dataproxy.ProxyClusterObject;
 import org.apache.inlong.commons.pojo.dataproxy.ProxySink;
 import org.apache.inlong.commons.pojo.dataproxy.ProxySource;
+import org.apache.inlong.commons.pojo.dataproxy.RepositoryTimerTask;
 import org.apache.inlong.manager.common.pojo.dataproxy.DataProxyClusterSet;
 import org.apache.inlong.manager.dao.entity.CacheCluster;
 import org.apache.inlong.manager.dao.entity.CacheClusterExt;
